+++ exec/plugin-commands-script-runners/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/plugin-commands-script-runners
 
+## 1000.0.2
+
+### Patch Changes
+
+- @pnpm/deps.status@1001.0.1
+- @pnpm/plugin-commands-installation@1001.0.1
+
 ## 1000.0.1
 
 ### Patch Changes