+++ lockfile/plugin-commands-audit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-audit
 
+## 1001.0.1
+
+### Patch Changes
+
+- Updated dependencies [3f0e4f0]
+  - @pnpm/lockfile.fs@1001.1.0
+  - @pnpm/audit@1001.0.1
+
 ## 1001.0.0
 
 ### Major Changes