+++ patching/plugin-commands-patching/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-patching
 
+## 1000.0.2
+
+### Patch Changes
+
+- Updated dependencies [3f0e4f0]
+  - @pnpm/lockfile.fs@1001.1.0
+  - @pnpm/plugin-commands-installation@1001.0.1
+
 ## 1000.0.1
 
 ### Patch Changes