+++ pkg-manager/plugin-commands-installation/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/plugin-commands-installation
 
+## 1001.0.1
+
+### Patch Changes
+
+- @pnpm/deps.status@1001.0.1
+- @pnpm/core@1001.0.1
+- @pnpm/get-context@1001.0.1
+- @pnpm/outdated@1001.0.1
+- @pnpm/plugin-commands-rebuild@1001.0.1
+
 ## 1001.0.0
 
 ### Major Changes