+++ pkg-manager/read-projects-context/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/read-projects-context
 
+## 1000.0.2
+
+### Patch Changes
+
+- Updated dependencies [3f0e4f0]
+  - @pnpm/lockfile.fs@1001.1.0
+
 ## 1000.0.1
 
 ### Patch Changes